@@ -11,101 +17,84 @@
   "publishConfig": {
     "access": "public"
   },
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
   "peerDependencies": {
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.13.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
-  "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/preset-env": "^7.16.4",
-    "@babel/preset-react": "^7.16.0",
-    "@babel/preset-typescript": "^7.16.0",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@mdx-js/react": "^2.1.2",
+  "dependencies": {
+    "@date-io/date-fns": "^3.0.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.13.2",
-    "@rollup/plugin-commonjs": "^25.0.0",
-    "@rollup/plugin-image": "^3.0.2",
-    "@rollup/plugin-node-resolve": "^15.1.0",
-    "@rollup/plugin-typescript": "^11.1.0",
-    "@storybook/addon-actions": "^7.0.18",
-    "@storybook/addon-essentials": "^7.0.18",
-    "@storybook/addon-links": "^7.0.18",
-    "@storybook/preset-scss": "^1.0.3",
-    "@storybook/react": "^7.0.18",
-    "@storybook/react-webpack5": "^7.0.18",
-    "@testing-library/react": "^14.0.0",
-    "@types/autosuggest-highlight": "^3.2.0",
-    "@types/jest": "^28.1.6",
-    "@types/react": "^18.2.7",
-    "@types/react-slick": "^0.23.10",
-    "@typescript-eslint/eslint-plugin": "^5.43.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "babel-jest": "^27.3.1",
-    "babel-loader": "^8.2.3",
-    "css-loader": "^6.5.1",
-    "eslint": "^8.44.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-config-standard-with-typescript": "^34.0.1",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-n": "^15.0.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^8.0.3",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^28.1.3",
-    "jest-environment-jsdom": "^29.5.0",
-    "lint-staged": "^13.2.2",
-    "postcss": "^8.4.23",
-    "prettier": "2.8.7",
+    "@mui/system": "^5.15.14",
+    "@mui/x-data-grid": "^6.19.8",
+    "@mui/x-date-pickers": "^6.19.8",
+    "autosuggest-highlight": "^3.3.4",
+    "date-fns": "^3.6.0",
+    "i18next": "^23.10.1",
+    "jquery": "^3.7.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "rollup": "^2.79.1",
-    "rollup-plugin-import-css": "^3.2.1",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-svg": "^2.0.0",
-    "sass": "^1.43.5",
-    "sass-loader": "^12.3.0",
-    "storybook": "^7.0.17",
-    "style-loader": "^3.3.1",
-    "typescript": "*",
-    "webpack": "^5.84.1"
-  },
-  "dependencies": {
-    "@mui/base": "^5.0.0-beta.3",
-    "@mui/system": "^5.13.2",
-    "@mui/x-data-grid": "^6.0.0",
-    "@mui/x-date-pickers": "^6.0.0",
-    "autosuggest-highlight": "^3.3.4",
-    "date-fns": "^2.30.0",
-    "i18next": "20.3.1",
-    "jquery": "^3.7.0",
-    "react-i18next": "^13.2.2",
-    "react-player": "^2.13.0",
-    "react-slick": "^0.29.0",
+    "react-i18next": "^14.1.0",
+    "react-player": "^2.15.1",
+    "react-slick": "^0.30.2",
     "slick-carousel": "^1.8.1"
   },
-  "resolutions": {
-    "@babel/traverse": "^7.23.2"
+  "devDependencies": {
+    "@babel/core": "^7.24.3",
+    "@babel/preset-react": "^7.24.1",
+    "@chromatic-com/storybook": "^1.2.25",
+    "@storybook/addon-essentials": "^8.0.2",
+    "@storybook/addon-interactions": "^8.0.2",
+    "@storybook/addon-links": "^8.0.2",
+    "@storybook/addon-onboarding": "^8.0.2",
+    "@storybook/blocks": "^8.0.2",
+    "@storybook/react": "^8.0.2",
+    "@storybook/react-vite": "^8.0.2",
+    "@storybook/test": "^8.0.2",
+    "@testing-library/react": "^14.2.2",
+    "@types/autosuggest-highlight": "^3.2.3",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.30",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
+    "@types/react-slick": "^0.23.13",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^8.57.0",
+    "eslint-config-love": "^43.1.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "eslint-plugin-storybook": "^0.8.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "prettier": "^3.2.5",
+    "sass": "^1.72.0",
+    "storybook": "^8.0.2",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "*",
+    "vite": "^5.1.6",
+    "vite-plugin-dts": "^3.7.3",
+    "vite-plugin-lib-inject-css": "^2.0.0"
   },
   "scripts": {
+    "build": "tsc --p ./tsconfig-build.json && vite build",
+    "lint": "eslint . --ext ts,tsx,js,jsx --report-unused-disable-directives --max-warnings 0",
     "pretty": "prettier --write \"**/*.{ts,tsx,js,jsx,json,css,sass,scss,xml,md}\"",
-    "lint": "yarn run eslint \"**/*.{ts,tsx,js,jsx}\"",
     "test": "jest",
-    "start": "storybook dev -p 3006",
-    "build": "rollup -c",
+    "test:ci": "CI=true jest",
+    "start:dev": "vite",
+    "start:storybook": "storybook dev -p 3006",
     "build:storybook": "storybook build -o ./storybook",
-    "build:sources": "zip -r portal-shared-components.zip src package.json yarn.lock -x '*.stories.*' -x '*.test.*' -x '*.css' -x '*.scss' -x '*.svg' -x '*.jpg' -x '*.png' -x '*.webp' -x '*.ttf'"
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   }
 }
