@@ -33,7 +33,7 @@ export default {
     progressPercent: 0,
     translations: {
       placeholder: 'Lorem Ipsum: Wieviele sachen willst du hochladen',
-      uploadProgess: 'Uploaded % of % files',
+      uploadProgress: 'Uploaded % of % files',
       uploadSuccess: 'Uploaded',
       uploadError: 'Not Uploaded',
     },
