@@ -46,12 +46,12 @@ export interface DropZoneDropAreaTranslations {
 
 export interface DropZonePreviewTranslations {
   placeholder: string | JSX.Element
-  uploadProgess: string
+  uploadProgress: string
   uploadSuccess: string | JSX.Element
   uploadError: string | JSX.Element
 }
 
-export interface deleteConfirmOverlayTranslation {
+export interface DeleteConfirmOverlayTranslation {
   title: string
   content: string
   action_no: string
