@@ -20,41 +20,39 @@
 
 export const Footer = () => {
   return (
-    <>
-      <svg
-        width="100%"
-        viewBox="0 0 1440 228"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-        style={{
-          width: '100% !important',
-          backgroundSize: '100% 100%',
-          bottom: 0,
-          position: 'absolute',
-        }}
-      >
-        <path
-          d="M719.5 141.724C479.667 172.04 239.833 172.513 0 143.141V228H1440V0C1383.59 11.9966 1327.18 24.5024 1270.77 37.0083L1270.48 37.0725C1086.82 77.7905 903.16 118.507 719.5 141.724Z"
-          fill="white"
-        />
-        <path
-          d="M719.5 141.724C479.667 172.04 239.833 172.513 0 143.141V228H1440V0C1383.59 11.9966 1327.18 24.5024 1270.77 37.0083L1270.48 37.0725C1086.82 77.7905 903.16 118.507 719.5 141.724Z"
-          fill="url(#paint0_radial_1011_87709)"
-        />
-        <defs>
-          <radialGradient
-            id="paint0_radial_1011_87709"
-            cx="0"
-            cy="0"
-            r="1"
-            gradientUnits="userSpaceOnUse"
-            gradientTransform="translate(698 200.305) rotate(-139.894) scale(305.939 890.489)"
-          >
-            <stop stopColor="#ADB9C7" />
-            <stop offset="1" stopColor="#F3F3F3" />
-          </radialGradient>
-        </defs>
-      </svg>
-    </>
+    <svg
+      width="100%"
+      viewBox="0 0 1440 228"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      style={{
+        width: '100% !important',
+        backgroundSize: '100% 100%',
+        bottom: 0,
+        position: 'absolute',
+      }}
+    >
+      <path
+        d="M719.5 141.724C479.667 172.04 239.833 172.513 0 143.141V228H1440V0C1383.59 11.9966 1327.18 24.5024 1270.77 37.0083L1270.48 37.0725C1086.82 77.7905 903.16 118.507 719.5 141.724Z"
+        fill="white"
+      />
+      <path
+        d="M719.5 141.724C479.667 172.04 239.833 172.513 0 143.141V228H1440V0C1383.59 11.9966 1327.18 24.5024 1270.77 37.0083L1270.48 37.0725C1086.82 77.7905 903.16 118.507 719.5 141.724Z"
+        fill="url(#paint0_radial_1011_87709)"
+      />
+      <defs>
+        <radialGradient
+          id="paint0_radial_1011_87709"
+          cx="0"
+          cy="0"
+          r="1"
+          gradientUnits="userSpaceOnUse"
+          gradientTransform="translate(698 200.305) rotate(-139.894) scale(305.939 890.489)"
+        >
+          <stop stopColor="#ADB9C7" />
+          <stop offset="1" stopColor="#F3F3F3" />
+        </radialGradient>
+      </defs>
+    </svg>
   )
 }
