@@ -85,7 +85,9 @@ export const MenuItem = ({
       }}
       onMouseEnter={onMouseEnter}
       onMouseLeave={onMouseLeave}
-      onClick={(e) => {(onClick == null || disable) ?? onClick(e)}}
+      onClick={(e) => {
+        ;(onClick == null || disable) ?? onClick(e)
+      }}
     >
       <Link
         component={component}
