@@ -67,6 +67,6 @@ Stepper.args = {
   list: stepperElements,
   showSteps: 6,
   activeStep: 3,
-  tooltipText: "Continue where you left",
-  tooltipLink: "/registration/form"
+  tooltipText: 'Continue where you left',
+  tooltipLink: '/registration/form',
 }
