@@ -48,7 +48,7 @@ function a11yProps(index: number) {
   }
 }
 
-function TabPanel(props: TabPanelProps) {
+function TabPanel(props: Readonly<TabPanelProps>) {
   const { children, value, index, ...other } = props
 
   return (
