@@ -28,9 +28,7 @@ export default {
 }
 
 const Template: ComponentStory<typeof Component> = (args: any) => (
-  <>
-    <Component {...args} />
-  </>
+  <Component {...args} />
 )
 
 const views: view[] = [
