@@ -43,13 +43,11 @@ interface CardsProps {
   imageShape?: CardProps['imageShape']
   imageLoader?: CardProps['imageLoader']
   imageElement?: CardProps['imageElement']
-  columns?: number
   readMoreText?: CardProps['readMoreText']
   readMoreLink?: CardProps['readMoreLink']
   addButtonClicked?: boolean
-  showAddNewCard?: boolean
   newButtonText?: string
-  onNewCardButton?: React.MouseEventHandler | undefined
+  onNewCardButton?: React.MouseEventHandler
   onCardClick?: (item: CardItems) => void
   subMenu?: boolean
   activeSubmenuOptions?: SubItems[]
@@ -71,11 +69,9 @@ export const Cards = ({
   imageSize,
   imageShape,
   imageLoader,
-  columns = 6,
   expandOnHover,
   filledBackground,
   addButtonClicked = false,
-  showAddNewCard = false,
   newButtonText,
   onNewCardButton,
   onCardClick = () => {
@@ -84,7 +80,7 @@ export const Cards = ({
   subMenu = false,
   activeSubmenuOptions = [],
   inactiveSubmenuOptions = [],
-  submenuClick = (sortMenu: string, id: string | undefined) => {
+  submenuClick = () => {
     // empty
   },
   tooltipText,
