@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2021, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021, 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -17,31 +17,6 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import React from 'react'
-import { SharedThemeProvider } from '../src/components/basic/SharedThemeProvider'
-import { SharedCssBaseline } from '../src/components/basic/SharedCssBaseline'
-import { I18nextProvider } from 'react-i18next'
-import i18n from '../src/i18next'
-
-const withI18next = (Story) => {
-  return (
-    <SharedThemeProvider>
-      <SharedCssBaseline />
-      <I18nextProvider i18n={i18n}>
-        <Story />
-      </I18nextProvider>
-    </SharedThemeProvider>
-  )
-}
-
-export const decorators = [withI18next]
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-}
+export * from './@types'
+export * from './components'
+export { theme, paletteDefinitions, typographyDefinitions } from './theme'
